// projects/wa-motif-open-api/app-content-service/src/lib/model/engineCreate.ts
/**
 * Motif App Content Service API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 2.0.0
 * Contact: info@vipera.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface EngineCreate { 
    /**
     * Engine name
     */
    name: string;
    /**
     * Engine latest version
     */
    latestVersion: string;
    /**
     * Forbidden versions
     */
    forbiddenVersion?: string;
    /**
     * Download URL
     */
    downloadUrl?: string;
}